set cancel order threshold"

+++ swp/orders/views.py
@@ -91,20 +91,30 @@ def store_in_items_list(student, items_purchased, items_post, creator):
                 obj1 = OrderList.objects.create(student=student, item=items_purchased[j], quantity=int(items_post[i]))
                 obj1.created_at=obj.timestamp
                 obj1.created_by=creator.username
+                obj1.modified_at=obj.timestamp
+                obj1.modified_by=creator.username
                 obj1.save()
             j += 1
 
 def delete_order(request, pk):
     obj = OrderList.objects.get(pk=pk)
-    obj2 = Items.objects.get(item_name=obj.item.item_name)
-    obj2.quantity += obj.quantity
-    obj2.save()
-    print(obj.item.item_name)
-    if obj:
-        obj1 = OrderHistory.objects.create(student=obj.student, item=obj.item, quantity=obj.quantity, timestamp=obj.timestamp)
-        obj1.save()
-        obj.delete()
-    return redirect('orders:orders_index')
+    current_date = datetime.datetime.now().date()
+    ordered_date = obj.created_at
+    delta_days = current_date.day - ordered_date.day
+
+    if current_date.year == ordered_date.year and current_date.month == ordered_date.month and abs(delta_days) <= 10:
+        print(delta_days)
+        obj2 = Items.objects.get(item_name=obj.item.item_name)
+        obj2.quantity += obj.quantity
+        obj2.save()
+        # print(obj.item.item_name)
+        if obj:
+            obj1 = OrderHistory.objects.create(student=obj.student, item=obj.item, quantity=obj.quantity, timestamp=obj.timestamp)
+            obj1.save()
+            obj.delete()
+        return redirect('orders:orders_index')
+    else:
+        return redirect('orders:orders_index')
 
 @login_required
 def getOrdersHistory(request):
@@ -139,7 +149,7 @@ def place_order(request):
         books = list(filter(lambda x: x.item_type == "Book", items))
         stationary = list(filter(lambda x: x.item_type == "Stationary", items))
         others = list(filter(lambda x: x.item_type == "Others", items))
-        print(request.POST)
+        # print(request.POST)
         books_post = request.POST.getlist('books')
         stationary_post = request.POST.getlist('stationary')
         others_post = request.POST.getlist('others')
@@ -156,7 +166,7 @@ def place_order(request):
             others_cost = dot([a.cost for a in others], list(map(int, others_post)))
 
         total_cost = books_cost + stationary_cost + others_cost
-        print(stationary_post)
+        # print(stationary_post)
         if books:
             books_purchased = [books[i] for i in range(len(books)) if books_post[i]!='0']
         if stationary:
@@ -182,20 +192,20 @@ def place_order(request):
             store_in_items_list(student, others_purchased, others_post, request.user)
             changeQuantity(others_purchased, others_post, not_order, ordered_items)
 
-        print('hi')
-        print(ordered_items)
-        print(not_order)
-        print('bye')
+        # print('hi')
+        # print(ordered_items)
+        # print(not_order)
+        # print('bye')
         if len(not_order) == 0 :
             return redirect('orders:orders_index')
         else:
             total_cost = 0
             for orders in ordered_items:
                 total_cost += orders.quantity*orders.cost
-            print(total_cost)
-            print('hi')
-            print(not_order)
-            print(ordered_items)
+            # print(total_cost)
+            # print('hi')
+            # print(not_order)
+            # print(ordered_items)
             return render(request, 'orders/not_order.html', {
             'ordered_items': ordered_items,
             'not_ordered': not_order,
